+++ src/decodeAsync.ts
@@ -1,16 +1,21 @@
 import { ExtensionCodecType } from "./ExtensionCodec";
 import { Decoder } from "./Decoder";
+import { defaultDecodeOptions, DecodeOptions } from "./decode";
 
-type DecodeAsyncOptions = Partial<
-  Readonly<{
-    extensionCodec: ExtensionCodecType;
-  }>
->;
+export type DecodeAsyncOptions = DecodeOptions;
+export const defaultDecodeAsyncOptions = defaultDecodeOptions;
 
 export async function decodeAsync(
   stream: AsyncIterable<Uint8Array | ArrayLike<number>>,
-  options?: DecodeAsyncOptions,
+  options: DecodeAsyncOptions = defaultDecodeOptions,
 ): Promise<unknown> {
-  const decoder = new Decoder(options && options.extensionCodec);
+  const decoder = new Decoder(
+    options.extensionCodec,
+    options.maxStrLength,
+    options.maxBinLength,
+    options.maxArrayLength,
+    options.maxMapLength,
+    options.maxExtLength,
+  );
   return decoder.decodeOneAsync(stream);
 }