+++ src/encode.ts
@@ -3,18 +3,16 @@ import { Encoder } from "./Encoder";
 
 export type EncodeOptions = Partial<
   Readonly<{
+    extensionCodec: ExtensionCodecType;
     maxDepth: number;
     initialBufferSize: number;
-    extensionCodec: ExtensionCodecType;
   }>
 >;
 
-export function encode(value: unknown, options?: EncodeOptions): Uint8Array {
-  const encoder = new Encoder(
-    options && options.extensionCodec,
-    options && options.maxDepth,
-    options && options.initialBufferSize,
-  );
+const defaultEncodeOptions = {};
+
+export function encode(value: unknown, options: EncodeOptions = defaultEncodeOptions): Uint8Array {
+  const encoder = new Encoder(options.extensionCodec, options.maxDepth, options.initialBufferSize);
   encoder.encode(value, 1);
   return encoder.getUint8Array();
 }