+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 This is the revision history of @msgpack/msgpack
 
+## v1.1.0 2019/05/19
+
+* Add options to `decode()` and `decodeAsync()`:
+  `maxStrLength`, `maxBinLength`, `maxArrayLength`, `maxMapLength`, and `maxExtLength` to limit max length of each item
+
 ## v1.0.1 2019/05/12
 
 * Fix IE11 incompatibility